mv index to src dir
