+++ package.json
@@ -2,7 +2,7 @@
   "name": "@enact/analytics",
   "version": "0.1.0",
   "description": "Generalized application interaction analytics library",
-  "main": "index.js",
+  "main": "src/index.js",
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1"
   },
